Renamed HomePost to Post and PostContainer
