+++ zuko-ui/src/pages/index.js
@@ -1,6 +1,6 @@
 import Stats from "../components/Stats/Stats";
 import QuickInformation from "../components/QuickInformation/QuickInformation";
-import HomePostsContainer from "../components/HomePostsContainer/HomePostsContainer";
+import PostsContainer from "../components/PostsContainer/PostsContainer";
 import StrategyGoals from "../components/StrategyGoals/StrategyGoals";
 import HeroBanner from '../components/HeroBanner/HeroBanner';
 
@@ -8,7 +8,7 @@ export default function HomePage() {
     return <div>
         <HeroBanner />
         <StrategyGoals />
-		<HomePostsContainer />
+		<PostsContainer />
         <Stats/>
         <QuickInformation />
     </div>
